Nico transactions save

+++ Back/src/main/java/com/mindhub/ticketmind/dtos/TicketFormDTO.java
@@ -4,5 +4,5 @@ import com.mindhub.ticketmind.models.TicketType;
 
 import java.math.BigDecimal;
 
-public record TicketFormDTO(String name, BigDecimal basePrice, int availableQuantity, String type) {
+public record TicketFormDTO(String name, double basePrice, int availableQuantity, String type) {
 }
